+++ R/daemon.R
@@ -107,7 +107,7 @@ daemon <- function(url, autoexit = TRUE, cleanup = TRUE, output = FALSE,
   cv <- cv()
   sock <- socket(protocol = "rep")
   on.exit(reap(sock))
-  autoexit && pipe_notify(sock, cv = cv, add = FALSE, remove = TRUE, flag = TRUE)
+  autoexit && pipe_notify(sock, cv = cv, remove = TRUE, flag = TRUE)
   if (length(tls)) tls <- tls_config(client = tls)
   dial_and_sync_socket(sock = sock, url = url, asyncdial = !autoexit, tls = tls)
 
@@ -146,13 +146,14 @@ daemon <- function(url, autoexit = TRUE, cleanup = TRUE, output = FALSE,
     count <- count + 1L
 
     (count >= maxtasks || count > timerstart && mclock() - start >= walltime) && {
+      next_config(mark = TRUE)
       send(ctx, data = data, mode = 3L)
-      data <- recv_aio_signal(sock, cv = cv, mode = 8L, timeout = .timelimit)
+      aio <- recv_aio_signal(ctx, cv = cv, mode = 8L)
       wait(cv)
       break
     }
 
-    send(ctx, data = data, mode = 1L)
+    send(ctx, data = data, mode = 3L)
     perform_cleanup(cleanup)
     if (count <= timerstart) start <- mclock()
 
@@ -187,16 +188,15 @@ daemon <- function(url, autoexit = TRUE, cleanup = TRUE, output = FALSE,
 
 dial_and_sync_socket <- function(sock, url, asyncdial, tls = NULL) {
   cv <- cv()
-  pipe_notify(sock, cv = cv, add = TRUE, remove = FALSE, flag = FALSE)
+  pipe_notify(sock, cv = cv, add = TRUE)
   dial(sock, url = url, autostart = asyncdial || NA, tls = tls, error = TRUE)
   wait(cv)
 }
 
-parse_cleanup <- function(cleanup) {
-  is.logical(cleanup) ||
-    return(c(as.integer(cleanup) %% 2L, (clr <- as.raw(cleanup)) & as.raw(2L), clr & as.raw(4L), clr & as.raw(8L)))
-  c(cleanup, cleanup, cleanup, FALSE)
-}
+parse_cleanup <- function(cleanup)
+  if (is.logical(cleanup))
+    c(cleanup, cleanup, cleanup, FALSE) else
+      c(as.integer(cleanup) %% 2L, (clr <- as.raw(cleanup)) & as.raw(2L), clr & as.raw(4L), clr & as.raw(8L))
 
 perform_cleanup <- function(cleanup) {
   if (cleanup[1L]) rm(list = (vars <- names(.GlobalEnv))[!vars %in% .[["vars"]]], envir = .GlobalEnv)